normal

+++ senpure-dispatcher-bean/src/main/java/com/senpure/dispatcher/criteria/ServerCenterAndWorkCriteria.java
@@ -9,13 +9,13 @@ import java.io.Serializable;
  * WorkId
  *
  * @author senpure-generator
- * @version 2019-3-12 9:58:25
+ * @version 2019-8-2 11:22:32
  */
 public class ServerCenterAndWorkCriteria extends Criteria implements Serializable {
     private static final long serialVersionUID = 1562592198L;
 
     //(主键)
-    private Long id;
+    private Integer id;
     //乐观锁，版本控制
     private Integer version;
     private String serverName;
@@ -101,7 +101,7 @@ public class ServerCenterAndWorkCriteria extends Criteria implements Serializabl
      *
      * @return
      */
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -110,7 +110,7 @@ public class ServerCenterAndWorkCriteria extends Criteria implements Serializabl
      *
      * @return
      */
-    public ServerCenterAndWorkCriteria setId(Long id) {
+    public ServerCenterAndWorkCriteria setId(Integer id) {
         this.id = id;
         return this;
     }
